+++ ispyb-ejb/src/test/java/ispyb/server/userportal/TestUserPortalUtils.java
